02_trading_bot/requirements.txt
python-binance
numpy
websocket_client